requirements/requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain_groq
faiss-cpu
faiss-gpu
python-dotenv
openai
tiktoken
kmeans_pytorch
recbole
sentence-transformers
python-dotenv
wikipedia